BFAC Location page
Added a BFAC location page for directions on how to get to The Center.
Changed the location link in about us page and hid history page utill
further notice.
Added BFAC Location page link in the main menu list and moved the EMF
links to the bottom of the menu.